+++ packages/elements/src/multi-input/index.ts
@@ -3,6 +3,7 @@ import {
   css,
   CSSResultGroup,
   html,
+  nothing,
   MultiValue,
   TemplateResult,
   PropertyValues
@@ -10,7 +11,6 @@ import {
 import { customElement } from '@refinitiv-ui/core/decorators/custom-element.js';
 import { property } from '@refinitiv-ui/core/decorators/property.js';
 import { createRef, ref, Ref } from '@refinitiv-ui/core/directives/ref.js';
-import { ifDefined } from '@refinitiv-ui/core/directives/if-defined.js';
 import { VERSION } from '../version.js';
 import { CollectionComposer } from '@refinitiv-ui/utils/collection.js';
 import type { Pill } from '../pill';
@@ -465,8 +465,8 @@ export class MultiInput extends ControlElement implements MultiValue {
           @keydown="${this.handleKeyDown}"
           @change="${this.onInputChange}"
           @input="${this.onInputChange}"
-          maxlength="${ifDefined(this.maxLength || undefined)}"
-          minlength="${ifDefined(this.minLength || undefined)}"
+          maxlength="${this.maxLength || nothing}"
+          minlength="${this.minLength || nothing}"
           ?error="${this.error}"
           .value="${this.value}"
           .icon="${this.icon || null}"